(#199) Fixes to support async initialization within tests for Cosmos.

+++ tests/CommunityToolkit.Datasync.TestCommon/Databases/CosmosDb/CosmosDbContext.cs
@@ -11,7 +11,7 @@ namespace CommunityToolkit.Datasync.TestCommon.Databases;
 [ExcludeFromCodeCoverage]
 public class CosmosDbContext(DbContextOptions<CosmosDbContext> options) : BaseDbContext<CosmosDbContext, CosmosEntityMovie>(options)
 {
-    public static CosmosDbContext CreateContext(string connectionString, ITestOutputHelper output = null, bool clearEntities = true)
+    public static async Task<CosmosDbContext> CreateContextAsync(string connectionString, ITestOutputHelper output = null, bool clearEntities = true)
     {
         if (string.IsNullOrEmpty(connectionString))
         {
@@ -23,20 +23,21 @@ public class CosmosDbContext(DbContextOptions<CosmosDbContext> options) : BaseDb
             .EnableLogging(output);
         CosmosDbContext context = new(optionsBuilder.Options);
 
-        context.InitializeDatabase(clearEntities);
-        context.PopulateDatabase();
+        await context.InitializeDatabaseAsync(clearEntities);
+        await context.PopulateDatabaseAsync();
         return context;
     }
 
-    internal void InitializeDatabase(bool clearEntities)
+    internal async Task InitializeDatabaseAsync(bool clearEntities)
     {
         if (clearEntities)
         {
-            // NOTE: sync-over-async is used here.  This is bad, but it is only used in the test suite.
-            // The test suite is not performance sensitive, so this is acceptable.
-            List<CosmosEntityMovie> movies = Movies.ToListAsync().Result;
-            RemoveRange(movies);
-            SaveChanges();
+            List<CosmosEntityMovie> movies = await Movies.ToListAsync();
+            if (movies.Count > 0)
+            {
+                RemoveRange(movies);
+                await SaveChangesAsync();
+            }
         }
     }
